Implement version-agnostic OSResponse.
Note 1: No "Base" in the class name because it is useable as-is.
Note 2: The OpenStack base classes I'm writing will be JSON-based. Not
wanting to rewrite the (soon-to-be-cut?) XML-based 1.0 driver, that will
instead inherit and override for XML where necessary, such as happens
here.
